requirements-account.txt
django-user-accounts==2.0.3
pinax-eventlog==2.0.3
pinax-webanalytics==4.0.2
django-bootstrap3==9.1.0
